Add functions to display the pipeline diagram on a window
One really helpful debugging aid is the ability to analyze graphically
the pipeline. This functions will allow our tracers to display in an
independent window the pipeline drawing. This was done by using
GStreamer's gst_debug_dot* capabilities along with graphviz graphs
layouts and renderers.

Along with a simple unit test, an example app was added to ease
integration with other modules.