Add a new thread function for displaying the pipeline graphic
A new function was added to the source code to create new thread of execution
when the function for graphic the pipeline is being call.

Because of the need that the pipeline graphic had to run on another thread so the
pipeline could continue running even when the window that displays the pipeline
graphic is close.

In order to do that the Glib libraries were used to create a new and independet
thread that receives as a parameter a structure with all the parameters included,
including a pointer to the graphic function.