Add a new plugin to display the pipeline graphic
The plugin GstGraphic was developed to obtain the graphic representation of the pipeline.

This plugin is needed to use the function available on the GstDot plugin, so when the
graphic plugin is being summon during the execution of the pipeline it will display the
pipeline graphic, and with this the user will get a way to inspect his pipeline easily.

All the logic inside the graphic plugin was developed with the purpose of using the
functions available on the GstDot plugin so the pipeline graphic can be displayed in a
new windows when the pipeline achieves the playing state.