Add API to compute the processing time for elements with one src pad and one sink pad
The processing time is measured creating an a list of all the elements
with one src pad and one sink pad. This list is used to identify
which element receives the buffer processed in the push pad event.
For each push pad event the processing time measured is the time between
the time where the buffer was processed by the previous element to the
time where the buffer was processed by the current element.