Add finalize function to GstGraphic plugin
A missing function on the GstGraphic plugin was added in order to
have the complete set of basic functions for the plugin.

When the pipeline is killed all the elements and plugins should
stop functioning, release allocated memory and close itself in
order to prevent future misfunctions because of residual process.

A finalize function was added to the source code with the instructions
for stopping the performance of the plugin correctly.