Add a new GstFramerate tracer plugin
A new plugin named GstFramerate that displays the frames per second for every source pad
on every element of pipeline was developed.

This plugin was developed in order to measure the framerate on every element of the pipeline,
so it will be easier for the user to inspect in which section of the pipeline is having a
decrease of the framerate.

All the logic and functions of the GstFramerate plugin were developed to measure how many
buffers go through every specific source pad of every element active on the pipeline and
displayed to the user when the pipeline achieves the playing state.