Add Interlatency tracer plugin
This plugin is part of the Gst-Shark tracers. Its function is
log the buffers latencies in nanoseconds, between the src and
all the intermediate elements pads in the pipeline.

When a buffer is sent from the source, a custom event is sent
downstream with the information to track the buffer and calculate
the time difference between the arrive time and the incoming time
of each buffer.

Interlatency is based in the latency gstreamer tracer plugin.