Add Scheduling time tracer
This tracer measures the time between one incoming buffer
and the next incoming buffer in a sink pad.
When a sink pad receives a buffer, the current timestamp is store
and when the next buffer is received the tracer compute the time
difference between the new timestamp and the previous
timestamp.