Add a API for storing every tracer results in a CTF file
Several functions were developed in order to correctly store the output of every tracer
in the Gst-Shark project in a Common Trace Format file.

This API was developed so the information displayed by every tracer can be stored in a
CTF file and therefore it can be analyzed after the execution of the pipeline. Also, this
output format matches the input expected by the Gst-Shark Eclipse plugin enabling the user
the possibility of using this tool as an interpreter of the information.

All the functions on the API were developed to store information of the results of every
executing tracer which means that there is a function specific for the output of every
tracer together with the corresponding functions to initialize and set the CTF files.