Add CTF print functions to every tracer of Gst-Shark
To every tracer available it was added a function to print its results to a common CTF
file, each tracer has its own specific format and it matches with the printing functions
of the API developed.

In order to correctly store the information deisplayed by every tracer at the time of
running the pipeline in a common CTF file, it was needed to develop and integrate the
logic needed to do so, respecting the rules of the Common Trace Format.

At the time of printing the results on the Debug display, each tracer also send this same
information to CTF API and it handle it to store the information in the correct format,
so it can be read after the execution has finalized with the Gst-Shark Eclipse plugin.