Add support for an environment variable that can disable the CTF output
An environment variable with which the generation of the CTF files can be disabled was defined,
so all the logic needed to handle the status of that variable was added in order of not affecting
the performance of the tracers even if the variable is set or not.

Depending on the type of tests that are going to be done with the tracers there would be sometimes
in which the CTF output files would not be necessary, so this logic was added in order to improve
the performance of the pipeline by removing the task of genertaing these files.

Based on the status of the environment variable there are some functions that will return after
checking the status of the global variable (linked to the environment variable) and without
executing creating or changing any of the CTF output files.