Add logic for definition of the evaluation version of Gst-Shark
All the logic needed to define an evaluation version with a restrictive
amount of execution time was added, this logic will not affect the
performance of the application on the complete version of it.

In order to provide a demo veriosn of application it eas needed to
limit the functionality of it by defining some limitations that will
only work fully if the complete version is purchased.

If the flag is correctly set at the time of configuring the package, it
will disable every output of every tracer after a defined and specific
amount of time.