Change all the variables name to GStreamer standard and add functionc to prevent memory leaks
Every variable name should respect a GStreamer standard so we can maintain the structure and format
of the tracers according to the GStreamer requirements. Also several for unreferencing and freeing
functions were added to the code.

Since this application is a library inside of GStreamer the format should be preserved, in order
to maintain the defined structure and make easier to following developers to debug the code. The
memory leaks prevention functions are for removing every remainder of the plugins after the
pipeline stops its execution.

All the variables names were changed according to the GStreamer standard and every function for
preventing a memory leak were added where it was needed, depending on the type of variables
and the functions implemented.