Add logic to enable/disable the CPUUsage tracer
A complete logic for checking the Operating System in which the application
is being run was added and depending on the result obtained Gst-Shark will
enable or disable the tracer for reading the CPU usage.

The method used for reading the load on the CPU is only valid for Linux
based systems and therefore the tracer crashes on other Operating Systems
and because of that it is needed to disable the tracer on those ones.

A logic to define a config variable was added to the configure.ac file so
at the time of building the application it will check for the Operating
System and enable or disable the config varable that will register the
tracer if the environment is suitable.