Add support for Gst-Shark on Gstreamer version 1.8.1
A functionality for checking the version of GStreamer in which Gst-Shark
is being installed was added to the configure.ac file and with that a
configuration variable was set depending on the result; based on that
variable the debug printing function on every tracer is selected.

In order to the application being able to work either with GStreamer 1.7.1,
GStreamer 1.8.1 or above it was needed to add the logic for checking the
version of GStreamer available at the time of building the plugin.

It was needed to add the proper function for printing the debug messages on
every tracer depending on the version of GStreamer, togeter with a selector
logic based on the results of the checks done previously at the time of
configuring the application.