Add support to create a serie of date for each field value for given field name
Filter all the event in the time range based in the name of the events.
Create a list of all the field values for the field name given.
For each field value, create a series of data.
Filter all the event in the time range based in the name field value.