Register remaining buffer flow hooks
In GStreamer, buffers may flow in various ways. The most common and intuitive way
is the push-mode scheduling. In this mode the upstream element pushes one buffer per
time to the downstream element. Additionally, GStreaemer supports pushing buffer lists,
and pull scheduling modes. We register a hook for each of the later.