CPU Usage: Fix invalid memory reads
During the tracer constructor a thread was spawned
computing CPU usage information every second. During
tracer destruction the thread attempted to read
object's members, which were no longer valid.

The thread approach was changed to a MainLoop
periodic timeout task, which in turn is removed
when the object is destroyed.