Process sink pads rather than source pads
Pad push hooks are installed by default in source pads. For the queue
level we are mostly interested in the sink pad, so we can report the new
levels as soon as a new buffer goes in, rather that as a buffer goes
out. To process the hook in the sink pads an intermediate function was
added that forwards it to the peer of the hooked pad.