fix plugin name
Recent versions of GStreamer are expecting to have the plugin name
matching the so file name.

With those versions, the plugin loader was looking for the
'gst_plugin_sharktracers_get_desc' symbol while 'gst_plugin_gstsharktracers_get_desc'
was defined.

Fix this by removing the 'gst' prefix in the plugin name.