Ensure one-time initialization of global log record
Even though GStreamer core only creates one instance of each tracer,
there is nothing that limits in the future the creation of multiple
instances. As such, it's better if we move the global log record
initialization to the class init, where we can ensure a one time call
even when multiple instances are created.