Implement refcounting mechanism for framerate activation
We only want to print framerate info if at least one pipe is running.
An application being profiled may consist of several pipelines and, as
such, we need to keep track of when the first pipeline starts and when
the last one stops, even though they may not be necessarily the same.