Fix timestamp minimum in plot_tracer when plotting traces
When a streamer client is started before a stream server, there was a time offset until the stream server started. This was caused by defining 0 as x min limit instead of computing the minimum timestamp of each trace. Thus, this fix changes the starting x limit from 0 to the minimum calculated in the timestamps of each trace.