Change plotting tool from default to gnuplot
This is part of a fix to deal with a bug when gstshark-plot have to plot multiple traces. Octave uses FLTK by default as plotting tool, but due to a bug in this library, it makes the gstShark-plot crashes when plot_tracer.m tries to plot multiple traces. This bug was also found in http://savannah.gnu.org/bugs/?45298, where this phenomenom happens when plotting with Octave as well. In order to fix this bug, the default plotting tool is substituted by GNU Plot.